Split project.py into two files
The file is way too very large which makes it annoying to work with.
It should also significantly speed up the pytype checks since right now
the majority of the time is spent analyzing project.py.
