+++ pycheribuild/projects/fvp_firmware.py
@@ -33,7 +33,8 @@ from pathlib import Path
 from .cross.crosscompileproject import CrossCompileMakefileProject
 from .cross.gdb import BuildGDB
 from .project import (DefaultInstallDir, GitRepository, MakefileProject, Project,
-                      ReuseOtherProjectDefaultTargetRepository, SimpleProject)
+                      ReuseOtherProjectDefaultTargetRepository)
+from .simple_project import SimpleProject
 from ..config.chericonfig import BuildType, CheriConfig
 from ..config.compilation_targets import CompilationTargets
 from ..config.loader import ComputedDefaultValue
