+++ pycheribuild/projects/kdevelop.py
@@ -27,7 +27,8 @@
 # OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
 # SUCH DAMAGE.
 #
-from .project import BuildType, CheriConfig, CMakeProject, DefaultInstallDir, GitRepository, SimpleProject
+from .project import BuildType, CheriConfig, CMakeProject, DefaultInstallDir, GitRepository
+from .simple_project import SimpleProject
 
 
 # doesn't seem to be part of distro packages
