+++ pycheribuild/projects/run_qemu.py
@@ -43,7 +43,8 @@ from .cross.gdb import BuildGDB
 from .cross.u_boot import BuildUBoot
 from .disk_image import (BuildCheriBSDDiskImage, BuildDiskImageBase, BuildFreeBSDImage,
                          BuildFreeBSDWithDefaultOptionsDiskImage, BuildMinimalCheriBSDDiskImage)
-from .project import CheriConfig, CPUArchitecture, SimpleProject, TargetAliasWithDependencies, Project
+from .project import CheriConfig, CPUArchitecture, Project
+from .simple_project import SimpleProject, TargetAliasWithDependencies
 from ..config.compilation_targets import CompilationTargets
 from ..config.loader import ComputedDefaultValue
 from ..qemu_utils import qemu_supports_9pfs, QemuOptions, riscv_bios_arguments
